Fixes #35: Hard code apollo dependency
