--- vue-cli-plugin-ic-apollo/index.js
@@ -1,2 +1,0 @@
-module.exports = () => {
-}
