+++ vue.config.js
@@ -12,16 +12,27 @@ const assetsPath = definePath(args.assetsPath, rootPath, config.assets)
 module.exports = {
   devServer: {
     before: configureAPI,
-    disableHostCheck: true,
+    disableHostCheck: true
   },
+
   chainWebpack: config => {
+    config.module
+      .rule('gql')
+      .test(/\.gql$/)
+      .use('graphql-tag/loader')
+      .loader('graphql-tag/loader')
+      .end()
+
     config
       .plugin('define')
       .tap(options => {
         options[0]['process.env'].COMPONENTS_BASEPATH = `"${componentsPath}"`
         options[0]['process.env'].PAGES_BASEPATH = `"${pagesPath}"`
+        options[0]['process.env'].IC_VUE_APP_GRAPHQL_WS = `"${process.env.IC_VUE_APP_GRAPHQL_WS || 'ws://localhost:4000/graphql'}"`
+        options[0]['process.env'].IC_VUE_APP_GRAPHQL_HTTP = `"${process.env.IC_VUE_APP_GRAPHQL_HTTP || 'http://localhost:4000/graphql'}"`
         return options
-      })
+      }
+      )
 
     config.resolve.alias
       .set('@components', componentsPath)
@@ -32,6 +43,7 @@ module.exports = {
     config.resolve.alias
       .set('@assets', assetsPath)
   },
+
   css: {
     sourceMap: true,
     loaderOptions: {
@@ -40,6 +52,13 @@ module.exports = {
           require('autoprefixer')()
         ]
       }
+    }
+  },
+
+  pluginOptions: {
+    apollo: {
+      enableMocks: true,
+      enableEngine: false
     }
   }
 }
